// apps/api/src/modules/home/domain/value-objects/home-section-enums.ts
export enum HomeSectionType {
  HERO_BANNER = 'HERO_BANNER',
  BANNER_GRID = 'BANNER_GRID',
  FEATURED_PRODUCTS = 'FEATURED_PRODUCTS',
  FEATURED_CATEGORIES = 'FEATURED_CATEGORIES',
  FEATURED_COLLECTIONS = 'FEATURED_COLLECTIONS',
  FEATURED_BRANDS = 'FEATURED_BRANDS',
  PROMOTION_BANNER = 'PROMOTION_BANNER',
  RICH_TEXT = 'RICH_TEXT',
  IMAGE_TEXT = 'IMAGE_TEXT',
  VIDEO_BANNER = 'VIDEO_BANNER',
  NEWSLETTER = 'NEWSLETTER',
}

export enum HomeSectionStatus {
  DRAFT = 'DRAFT',
  PUBLISHED = 'PUBLISHED',
}